+++ BovineLabs.Core.PerformanceTests/AssemblyInfo.cs
@@ -1,7 +1,0 @@
-// <copyright file="AssemblyInfo.cs" company="BovineLabs">
-//     Copyright (c) BovineLabs. All rights reserved.
-// </copyright>
-
-using Unity.Entities;
-
-[assembly: DisableAutoCreation]