+++ BovineLabs.Core.Tests/Jobs/IJobHashMapDeferTests.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="IJobHashMapVisitKeyValueTests.cs" company="BovineLabs">
+﻿// <copyright file="IJobHashMapDeferTests.cs" company="BovineLabs">
 //     Copyright (c) BovineLabs. All rights reserved.
 // </copyright>