+++ BovineLabs.Core/Extensions/UnsafeParallelHashMapExtensions.cs
@@ -6,12 +6,64 @@ namespace BovineLabs.Core.Extensions
 {
     using System;
     using System.Diagnostics;
+    using System.Threading;
     using Unity.Burst;
     using Unity.Collections;
     using Unity.Collections.LowLevel.Unsafe;
 
-    public static class UnsafeParallelHashMapExtensions
+    public static unsafe class UnsafeParallelHashMapExtensions
     {
+        public static ref TValue GetOrAddRef<TKey, TValue>(this UnsafeParallelHashMap<TKey, TValue>.ParallelWriter hashMap, TKey key, TValue defaultValue = default)
+            where TKey : unmanaged, IEquatable<TKey>
+            where TValue : unmanaged
+        {
+            var data = hashMap.m_Buffer;
+
+            if (UnsafeParallelHashMapBase<TKey, TValue>.TryGetFirstValueAtomic(data, key, out _, out var tempIt))
+            {
+                return ref UnsafeUtility.ArrayElementAsRef<TValue>(data->values, tempIt.EntryIndex);
+            }
+
+            // Allocate an entry from the free list
+            int idx = UnsafeParallelHashMapBase<TKey, TValue>.AllocEntry(data, hashMap.ThreadIndex);
+
+            // Write the new value to the entry
+            UnsafeUtility.WriteArrayElement(data->keys, idx, key);
+            UnsafeUtility.WriteArrayElement(data->values, idx, defaultValue);
+
+            int bucket = key.GetHashCode() & data->bucketCapacityMask;
+            // Add the index to the hash-map
+            int* buckets = (int*)data->buckets;
+
+            // Make the bucket's head idx. If the exchange returns something other than -1, then the bucket had
+            // a non-null head which means we need to do more checks...
+            if (Interlocked.CompareExchange(ref buckets[bucket], idx, -1) != -1)
+            {
+                int* nextPtrs = (int*)data->next;
+                int next;
+
+                do
+                {
+                    // Link up this entry with the rest of the bucket under the assumption that this key
+                    // doesn't already exist in the bucket. This assumption could be wrong, which will be
+                    // checked later.
+                    next = buckets[bucket];
+                    nextPtrs[idx] = next;
+
+                    // If the key already exists then we should free the entry we took earlier.
+                    if (UnsafeParallelHashMapBase<TKey, TValue>.TryGetFirstValueAtomic(data, key, out _, out tempIt))
+                    {
+                        // Put back the entry in the free list if someone else added it while trying to add
+                        UnsafeParallelHashMapBase<TKey, TValue>.FreeEntry(data, idx, hashMap.ThreadIndex);
+                        return ref UnsafeUtility.ArrayElementAsRef<TValue>(data->values, tempIt.EntryIndex);
+                    }
+                }
+                while (Interlocked.CompareExchange(ref buckets[bucket], idx, next) != next);
+            }
+
+            return ref UnsafeUtility.ArrayElementAsRef<TValue>(data->values, idx);
+        }
+
         public static unsafe void ClearAndAddBatchUnsafe<TKey, TValue>(
             [NoAlias] ref this UnsafeParallelHashMap<TKey, TValue> hashMap,
             [NoAlias] TKey* keys,